requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2
httpx[http2]==0.25.2
aiofiles==23.2.1
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydub==0.25.1
typing_extensions==4.14.1
cachetools==6.1.0
elastic-apm==6.24.0
pinecone==7.3.0
openai==1.104.2
# Logging
loguru==0.7.2
# Optional: For better performance
orjson==3.9.10
